requirements.txt
bcrypt==3.1.3
cffi==1.10.0
Django==1.11.4
pycparser==2.18
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
